backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
cryptography>=39.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-generativeai>=0.8.0
celery>=5.3.0
redis>=4.5.0
pydantic>=2.0.0
requests>=2.28.0
